Refactor text animation implementation by removing TextAnimator and TextAnimationDirection; introduce GraphicsTextAnimator for improved functionality
