--- src/main/java/com/pi4j/drivers/display/graphics/text/TextAnimationDirection.java
@@ -1,6 +1,0 @@
-package com.pi4j.drivers.display.graphics.text;
-
-public enum TextAnimationDirection {
-    RIGHT_TO_LEFT,
-    LEFT_TO_RIGHT
-}
